Do some file location refactoring
